refactor: move directive codegen helpers
