+++ crates/vue_oxlint_jsx/src/parser/codegen.rs
@@ -1,4 +1,4 @@
-use oxc_ast::ast::JSXChild;
+use oxc_ast::ast::{JSXChild, Str};
 use oxc_span::Span;
 
 use crate::parser::ParserImpl;
@@ -27,4 +27,31 @@ impl<'a> ParserImpl<'a> {
 
     self.ast.jsx_child_text(span, ast_str, Some(ast_str))
   }
+
+  pub(super) fn codegen_directive_identifier(&self, name: &'a str) -> Str<'a> {
+    if !self.config.codegen || is_codegen_safe_jsx_identifier(name) {
+      return name.into();
+    }
+
+    let mut result = String::from("__v_");
+    for ch in name.chars() {
+      if ch.is_ascii_alphanumeric() || matches!(ch, '_' | '$' | '-') {
+        result.push(ch);
+      } else if !result.ends_with('_') {
+        result.push('_');
+      }
+    }
+
+    result.push_str("__");
+    self.ast.str(&result)
+  }
+}
+
+fn is_codegen_safe_jsx_identifier(name: &str) -> bool {
+  let Some((&first, rest)) = name.as_bytes().split_first() else {
+    return false;
+  };
+
+  matches!(first, b'a'..=b'z' | b'A'..=b'Z' | b'_' | b'$')
+    && rest.iter().all(|b| b.is_ascii_alphanumeric() || matches!(b, b'_' | b'$' | b'-'))
 }
