+++ crates/vue_oxlint_jsx/src/parser/elements/directive.rs
@@ -1,5 +1,5 @@
 use memchr::memchr;
-use oxc_ast::ast::{JSXAttributeName, Str};
+use oxc_ast::ast::JSXAttributeName;
 use oxc_span::{SPAN, Span};
 use vue_compiler_core::parser::Directive;
 
@@ -67,41 +67,6 @@ impl<'a> ParserImpl<'a> {
       ),
     )
   }
-
-  fn codegen_directive_identifier(&self, name: &'a str) -> Str<'a> {
-    if !self.config.codegen || is_codegen_safe_jsx_identifier(name) {
-      return name.into();
-    }
-
-    let mut result = String::from("__v_");
-    for ch in name.chars() {
-      if ch.is_ascii_alphanumeric() || matches!(ch, '_' | '$' | '-') {
-        result.push(ch);
-      } else if !result.ends_with('_') {
-        result.push('_');
-      }
-    }
-
-    if result == "__v_" {
-      return self.ast.str("__v__");
-    }
-
-    if result.ends_with('_') {
-      result.push('_');
-    } else {
-      result.push_str("__");
-    }
-    self.ast.str(&result)
-  }
-}
-
-fn is_codegen_safe_jsx_identifier(name: &str) -> bool {
-  let Some((&first, rest)) = name.as_bytes().split_first() else {
-    return false;
-  };
-
-  matches!(first, b'a'..=b'z' | b'A'..=b'Z' | b'_' | b'$')
-    && rest.iter().all(|b| b.is_ascii_alphanumeric() || matches!(b, b'_' | b'$' | b'-'))
 }
 
 #[cfg(test)]
